Added useBetActions custom hook
